auth_service/requirements.txt
anyio==3.6.2 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.10" and python_version < "4.0"
attrs==22.1.0 ; python_version >= "3.10" and python_version < "4.0"
backoff==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "4"
cffi==1.15.1 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==2.1.1 ; python_version >= "3.10" and python_version < "4"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0"
cryptography==39.0.1 ; python_version >= "3.10" and python_version < "4.0"
curlify==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.1.0 ; python_version >= "3.10" and python_version < "3.11"
flask-admin==1.6.1 ; python_version >= "3.10" and python_version < "4.0"
flask-jwt-extended==4.4.4 ; python_version >= "3.10" and python_version < "4"
flask-peewee==3.0.4 ; python_version >= "3.10" and python_version < "4.0"
flask-wtf==1.1.1 ; python_version >= "3.10" and python_version < "4.0"
flask==2.2.3 ; python_version >= "3.10" and python_version < "4.0"
gevent==22.10.2 ; python_version >= "3.10" and python_version < "4.0"
graphql-core==3.2.3 ; python_version >= "3.10" and python_version < "4"
greenlet==2.0.2 ; python_version >= "3.10" and python_version < "4.0" and platform_machine == "aarch64" or python_version >= "3.10" and python_version < "4.0" and platform_machine == "ppc64le" or python_version >= "3.10" and python_version < "4.0" and platform_machine == "x86_64" or python_version >= "3.10" and python_version < "4.0" and platform_machine == "amd64" or python_version >= "3.10" and python_version < "4.0" and platform_machine == "AMD64" or python_version >= "3.10" and python_version < "4.0" and platform_machine == "win32" or python_version >= "3.10" and python_version < "4.0" and platform_machine == "WIN32" or python_version >= "3.10" and python_version < "4.0" and platform_python_implementation == "CPython"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==0.16.3 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.23.3 ; python_version >= "3.10" and python_version < "4.0"
hypothesis-graphql==0.9.2 ; python_version >= "3.10" and python_version < "4.0"
hypothesis-jsonschema==0.22.1 ; python_version >= "3.10" and python_version < "4.0"
hypothesis==6.68.2 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4"
iniconfig==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
jsonschema==4.17.3 ; python_version >= "3.10" and python_version < "4.0"
junit-xml==1.9 ; python_version >= "3.10" and python_version < "4.0"
loguru==0.6.0 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.0 ; python_version >= "3.10" and python_version < "4.0"
peewee-migrate==1.6.6 ; python_version >= "3.10" and python_version < "4.0"
peewee==3.15.4 ; python_version >= "3.10" and python_version < "4.0"
pluggy==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.5 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.5 ; python_version >= "3.10" and python_version < "4.0"
pyjwt==2.6.0 ; python_version >= "3.10" and python_version < "4"
pyrsistent==0.19.3 ; python_version >= "3.10" and python_version < "4.0"
pytest-subtests==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
pytest==7.2.1 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==0.21.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
redis==4.5.1 ; python_version >= "3.10" and python_version < "4.0"
requests==2.28.1 ; python_version >= "3.10" and python_version < "4"
rfc3986[idna2008]==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
schemathesis==3.18.5 ; python_version >= "3.10" and python_version < "4.0"
setuptools==67.4.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
sortedcontainers==2.4.0 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==2.0.4 ; python_version >= "3.10" and python_version < "4.0"
starlette-testclient==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.25.0 ; python_version >= "3.10" and python_version < "4.0"
tomli-w==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
types-peewee==3.15.0.7 ; python_version >= "3.10" and python_version < "4.0"
types-psycopg2==2.9.21.6 ; python_version >= "3.10" and python_version < "4.0"
types-pyopenssl==23.0.0.4 ; python_version >= "3.10" and python_version < "4.0"
types-redis==4.5.1.3 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.14 ; python_version >= "3.10" and python_version < "4"
werkzeug==2.2.3 ; python_version >= "3.10" and python_version < "4.0"
wheel==0.38.1 ; python_version >= "3.10" and python_version < "4.0"
win32-setctime==1.1.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
wtf-peewee==3.0.4 ; python_version >= "3.10" and python_version < "4.0"
wtforms==3.0.1 ; python_version >= "3.10" and python_version < "4.0"
yarl==1.8.2 ; python_version >= "3.10" and python_version < "4.0"
zope-event==4.6 ; python_version >= "3.10" and python_version < "4.0"
zope-interface==5.5.2 ; python_version >= "3.10" and python_version < "4.0"
